image input working
rough working version of image support. refinement to come later.
